Created menu_handler_module and moved exit_app method
